refactor: add new logging backend and clean up old action panels

+++ menu/power/EntryPowerOff.java
@@ -5,6 +5,8 @@ import java.io.IOException;
 
 import menu.MenuEntry;
 import menu.MenuOptionList;
+import system.Config;
+import util.Log;
 
 class EntryPowerOff extends MenuEntry {
 
@@ -14,10 +16,14 @@ class EntryPowerOff extends MenuEntry {
 
     @Override
     public void execute() {
-        System.out.println("Need to execute \"systemctl poweroff\"");
+        if (!Config.ENABLE_POWER_MANAGEMENT) {
+            Log.logInfo("menu.pm: power management requested but disabled");
+            return;
+        }
 
         try {
             Runtime.getRuntime().exec(new String[] {"systemctl", "poweroff"});
+            Log.logVerbose("menu.pm: executing \"systemctl poweroff\"");
         }
         catch (IOException ie) {
             //.. doesn't really matter?
